albedo normal buffer half3

+++ zenovis/xinxinoptix/optixPathTracer.h
@@ -163,8 +163,8 @@ struct Params
     float3*      frame_buffer_M;
 
     float3*      debug_buffer;
-    float3*      albedo_buffer;
-    float3*      normal_buffer;
+    ushort3*     albedo_buffer;
+    ushort3*     normal_buffer;
 
     unsigned int width;
     unsigned int height;
